+++ cpp/search/search.cpp
@@ -1078,7 +1078,7 @@ void Search::addDirichletNoise(const SearchParams& searchParams, Rand& rand, int
   for(int i = 0; i<policySize; i++) {
     if(policyProbs[i] >= 0) {
       double weight = searchParams.rootDirichletNoiseWeight;
-      policyProbs[i] = r[i] * weight + policyProbs[i] * (1.0-weight);
+      policyProbs[i] = (float)(r[i] * weight + policyProbs[i] * (1.0-weight));
     }
   }
 }
@@ -1112,7 +1112,7 @@ void Search::maybeAddPolicyNoise(SearchThread& thread, SearchNode& node, bool is
     for(int i = 0; i<policySize; i++) {
       if(node.nnOutput->policyProbs[i] > 0) {
         //Numerically stable way to raise to power and normalize
-        double p = exp((log((double)node.nnOutput->policyProbs[i]) - logMaxValue) * invTemp);
+        float p = (float)exp((log((double)node.nnOutput->policyProbs[i]) - logMaxValue) * invTemp);
         node.nnOutput->policyProbs[i] = p;
         sum += p;
       }
@@ -1120,7 +1120,7 @@ void Search::maybeAddPolicyNoise(SearchThread& thread, SearchNode& node, bool is
     assert(sum > 0.0);
     for(int i = 0; i<policySize; i++) {
       if(node.nnOutput->policyProbs[i] >= 0) {
-        node.nnOutput->policyProbs[i] = (double)node.nnOutput->policyProbs[i] / sum;
+        node.nnOutput->policyProbs[i] = (float)(node.nnOutput->policyProbs[i] / sum);
       }
     }
   }
