+++ cpp/tests/testtrainingwrite.cpp
@@ -24,7 +24,7 @@ static NNEvaluator* startNNEval(
   int nnMutexPoolSizePowerOfTwo = 12;
   bool debugSkipNeuralNet = modelFile == "/dev/null";
   bool alwaysIncludeOwnerMap = false;
-  double nnPolicyTemperature = 1.0;
+  float nnPolicyTemperature = 1.0;
   const string openCLTunerFile = "";
   NNEvaluator* nnEval = new NNEvaluator(
     modelName,
@@ -204,7 +204,7 @@ void Tests::runSelfplayInitTestsWithNN(const string& modelFile) {
     fancyModes.forkSidePositionProb = 0.40;
     fancyModes.cheapSearchProb = 0.5;
     fancyModes.cheapSearchVisits = 20;
-    fancyModes.cheapSearchTargetWeight = 0.123;
+    fancyModes.cheapSearchTargetWeight = 0.123f;
     fancyModes.earlyForkGameProb = 0.5;
     fancyModes.earlyForkGameExpectedMoveProp = 0.05;
     fancyModes.earlyForkGameMinChoices = 2;
